ml_scripts/requirements.txt
scikit-learn==1.5.0
xgboost==2.0.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
transformers>=4.30.0
torch>=2.0.0
fpdf2>=2.7.5
sentencepiece>=0.1.99
